Добавляет модуляцию в проект
